+++ Assets/Scripts/GameLogic/Entities/Player.cs
@@ -8,6 +8,7 @@ namespace Ventura.GameLogic.Entities
     {
         public Player(string name) : base(name)
         {
+            _spriteId = "player";
             _inventory = new Inventory(this, 999);
             _skills = new Skills(this);
         }