+++ Assets/Scripts/Generators/PlayerGenerator.cs
@@ -1,4 +1,5 @@
 using Ventura.GameLogic.Entities;
+using Ventura.Util;
 
 namespace Ventura.Generators
 {
@@ -8,10 +9,10 @@ namespace Ventura.Generators
 
         public static Player GeneratePlayerWithBooks()
         {
-            //DebugUtils.Log("GeneratePlayer()");
+            DebugUtils.Log("GeneratePlayer()");
             var player = new Player("AnGian");
-            foreach (var book in BookItemGenerator.Instance.GenerateBooks())
-                player.Inventory.AddItem(book);
+            //foreach (var book in BookItemGenerator.Instance.GenerateBooks())
+            //    player.Inventory.AddItem(book);
 
             return player;
         }